programs/notion-framework/pyproject.toml
[project]
name = "notion-framework"
version = "0.1.0"
description = "A framework for generating typed tools from Notion database schemas"
authors = [{ name = "Nathan Luo", email = "nathanluo13@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "llmgine",
    "notion-client>=2.3.0",
    "pydantic>=2.11.3",
    "jinja2>=3.1.0",
    "aiofiles>=24.1.0",
    "click>=8.0.0",
    "rich>=13.9.4",
    "python-dotenv>=1.1.0",
]

[project.scripts]
notion-framework = "notion_framework.cli:cli"

[tool.uv.sources]
llmgine = { path = "../../", editable = true }

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.10.0",
    "mypy>=0.991",
    "ruff>=0.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/notion_framework"]

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-sv --log-cli-level=0"

[tool.ruff]
target-version = "py39"
line-length = 90
fix = true

[tool.ruff.lint]
select = [
    "YTT", "S", "B", "A", "C4", "T10", "SIM", "I", "C90", "E", "W", "F", "PGH", "RUF"
]
ignore = ["E501", "E731", "W293"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true